chore: bump to 0.14.0 — item 19 complete (19/19 done)
Split devtools/manager.ts (867 lines) into 3 modules:
- network-capture.ts (165 lines): network request/response ring buffer
- page-inspector.ts (250 lines): DOM/storage/performance/screenshot queries
- manager.ts (525 lines): CDP lifecycle coordinator + delegation facade

Composition + delegation pattern (same as item #18). Zero consumer changes.
