+++ docs/code-quality/STATUS.md
@@ -5,9 +5,9 @@
 
 ## Current State
 
-**Version:** 0.13.0
-**Last completed item:** #18
-**Overall:** 18/19 done
+**Version:** 0.14.0
+**Last completed item:** #19
+**Overall:** 19/19 done ✅
 
 ---
 
@@ -43,13 +43,20 @@
 |---|-------------|--------|---------|--------|
 | 17 | **API route tests** — Add integration tests for all 12 route files (~3000 lines total). Needs Express mocking setup | DONE | 2026-02-27 | 20ffdf3 |
 | 18 | **Split security-db.ts** — Split 958-line file by table group into `db-events.ts`, `db-baselines.ts`, `db-blocklist.ts` | DONE | 2026-02-27 | 351f7e2 |
-| 19 | **Split devtools/manager.ts** — Split 863-line file into CDP lifecycle manager + storage/DOM/performance inspector | TODO | | |
+| 19 | **Split devtools/manager.ts** — Split 863-line file into CDP lifecycle manager + storage/DOM/performance inspector | DONE | 2026-02-27 | be9be7f |
 
 ---
 
 ## Session Log
 
 <!-- Add an entry after each session -->
+
+### 2026-02-27 — Session 6: Item 19 (Split devtools/manager.ts)
+
+- **Items completed:** #19
+- **Version bumped to:** 0.14.0
+- **Commit(s):** `be9be7f`
+- **Notes:** Split 867-line `DevToolsManager` into 3 modules using composition + delegation pattern. `network-capture.ts` (165 lines: network request/response ring buffer + filtering), `page-inspector.ts` (250 lines: DOM/storage/performance/screenshot queries). `manager.ts` reduced to 525 lines (CDP lifecycle coordinator + subscriber system + copilot vision + delegation). Zero changes to 10 consumer files. 940 tests passing, 0 TS errors. All 19/19 code quality items complete.
 
 ### 2026-02-27 — Session 5: Item 18 (Split security-db.ts)
