+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tandem-browser",
-  "version": "0.13.0",
+  "version": "0.14.0",
   "description": "AI-Human symbiotic browser — Kees & Robin browse as one",
   "main": "dist/main.js",
   "author": "Robin Waslander <keesbeest78@proton.me>",
